Add CSV writer interface
- SkyDolly CSV writer implemenation
