+++ src/Plugin/src/Export/CsvExport/CMakeLists.txt
@@ -18,6 +18,9 @@ target_sources(${LIBRARY_NAME}
         src/CsvExportPlugin.h src/CsvExportPlugin.cpp
         src/CsvExportSettings.h src/CsvExportSettings.cpp
         src/CsvExportOptionWidget.h src/CsvExportOptionWidget.cpp src/CsvExportOptionWidget.ui
+        src/CsvWriterIntf.h
+        src/SkyDollyCsvWriter.h src/SkyDollyCsvWriter.cpp
+        src/FlightRadar24CsvWriter.h src/FlightRadar24CsvWriter.cpp
         src/CsvExportPlugin.json
 )
